% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert Year, Month, and Day Columns to a Single Date Object}
\usage{
convert_to_date(
  data,
  year_col = "iyear",
  month_col = "imonth",
  day_col = "iday"
)
}
\arguments{
\item{data}{The data frame containing the separate year, month, and day columns.}

\item{year_col}{Name of the column representing the year. Default is "iyear".}

\item{month_col}{Name of the column representing the month. Default is "imonth".}

\item{day_col}{Name of the column representing the day. Default is "iday".}
}
\value{
A single date object representing the combined year, month, and day.
}
\description{
This function takes a data frame with separate columns for year, month, and day
and combines them into a single date object.
}
\examples{
# Load your data
gta.df <- read.csv(system.file("data", "FTerrorData.csv", package = "GlobalTerrorAnalyticsGTA"))

# Convert separate year, month, and day columns to a single date column
gta.df$date <- convert_to_date(gta.df)

}
