% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3.R
\name{count_bombing_explosions}
\alias{count_bombing_explosions}
\title{Count the number of terrorist incidents with attack type "Bombing/Explosion"}
\usage{
count_bombing_explosions(data)
}
\arguments{
\item{data}{A data frame containing the terrorist incident data.}
}
\value{
The number of terrorist incidents with the attack type "Bombing/Explosion"
}
\description{
Count the number of terrorist incidents with attack type "Bombing/Explosion"
}
\examples{
# Load your data
gta.df <- read.csv(system.file("data", "FTerrorData.csv", package = "GlobalTerrorAnalyticsGTA"))

# Count the number of terrorist incidents with attack type "Bombing/Explosion"
count_bombing_explosions(gta.df)
}
