% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function5.R
\name{plot_attacks}
\alias{plot_attacks}
\title{Plot Type of Attacks}
\usage{
plot_attacks(data)
}
\arguments{
\item{data}{A dataframe containing the data.}
}
\value{
A ggplot object.
}
\description{
Plot the number of attacks over time grouped by attack type.
}
\examples{
data(gta.df)
plot_attacks(gta.df)
}
