% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function4.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate Summary Statistics}
\usage{
summary_stats(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named vector containing the mean, median, and standard deviation.
}
\description{
Calculate mean, median, and standard deviation for a numeric vector.
}
\examples{
# Load your data
gta.df <- read.csv(system.file("data", "FTerrorData.csv", package = "GlobalTerrorAnalyticsGTA"))

# Calculate summary statistics for the number of fatalities (nkill)
nkill_stats <- summary_stats(gta.df$nkill)
cat("nkill statistics:", nkill_stats)
}
