% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{visualize_timeseries}
\alias{visualize_timeseries}
\title{Creates a timeseries plot showing number of terrorist incidents for each year}
\usage{
visualize_timeseries(data)
}
\arguments{
\item{data}{A data frame containing the terrorist incident data.}
}
\value{
A ggplot2 plot object
}
\description{
Creates a timeseries plot showing number of terrorist incidents for each year
}
\details{
This function takes a data frame containing information about terrorist incidents
and creates a timeseries plot showing the number of incidents for each year.
}
\examples{
# Load your data
gta.df <- read.csv(system.file("data", "FTerrorData.csv", package = "GlobalTerrorAnalyticsGTA"))

# Visualize timeseries
visualize_timeseries(gta.df)

}
